fix(platform): keep heavy formatters out of the completion fast path
Extracting renderPlatformGrid() into platforms.ts pulled static imports
of formatters/markdown.js, formatters/table.js, and formatters/text-table.js
into a module that complete.ts's shell-completion fast path statically
imports for VALID_PLATFORMS/COMMON_PLATFORMS. Every `__complete`
invocation was transitively loading chalk/marked/cli-highlight/
string-width/wrap-ansi — deps that fast path is explicitly designed to
avoid (per its own file header: "no Stricli boot ... minimal startup
time"). Flagged by Bugbot.

Move renderPlatformGrid() to a new lib/platform-grid.ts that only
project/create.ts and platform/list.ts import; platforms.ts goes back
to importing just fuzzy.js's levenshtein, matching its pre-#1366 shape.
