+++ .lore.md
@@ -194,6 +194,9 @@
 <!-- lore:019ef4c9-4cc4-78d0-9f45-fdd69b883ae5 -->
 * **bundle-sources.ts exit code 1 for no-sources relies on cli.ts never resetting exitCode — fragile invariant**: Trap: using \`this.process.exitCode = 1\` directly (at \`bundle-sources.ts:145\`) instead of throwing \`OutputError\`(=60) looks like a valid shortcut. It works today because \`cli.ts:622-649\` never resets \`exitCode\` on a clean return. But this is a fragile invariant — if \`cli.ts\` ever resets \`exitCode\`, the no-sources case silently exits 0. The \`OutputError\`(=60) idiom is the correct pattern. The \`exitCode=1\` approach was kept for consistency with \`check.ts\` pattern; a comment was added explaining the choice. Consider migrating to \`OutputError\` in a future cleanup.
 
+<!-- lore:019fd66d-5e71-7521-b979-62be38cef48b -->
+* **check:fragments fails when new command has no fragment file in apps/cli-docs/src/fragments/commands/**: Trap: adding a new top-level command (e.g. \`platform\`) without a matching fragment file in \`apps/cli-docs/src/fragments/commands/\` causes \`check:fragments\` (packages/cli/script/check-fragments.ts) to exit 1 in CI's 'Validate generated files' job. Looks like a docs-only concern but blocks CI. Fix: create \`apps/cli-docs/src/fragments/commands/\<command>.md\` with at least one bash example block covering each subcommand. The script expects 34 routes + 1 index = 35 total fragment files; it also warns (non-blocking without \`--strict\`) when a fragment doesn't mention a subcommand. Pre-existing warning: \`proguard.md\` missing \`upload\` subcommand coverage — non-blocking.
+
 <!-- lore:019ef465-caa1-7e83-aa8b-4887246a1d55 -->
 * **check.ts hasId() uses !== null but ObjectFile.codeId is string|undefined — null guard mismatch**: Trap: \`check.ts\` \`hasId()\` guards \`o.codeId !== null && o.codeId.length > 0\`. Looks correct because old \`DifObjectInfo.codeId\` was \`string | null\`. But \`ObjectFile.codeId\` in \`@sentry/symbolic\` 13.4.0 is \`string | undefined\` — the getter returns \`undefined\` (not \`null\`) when Rust returns \`None\`. If \`DifObjectInfo.codeId\` is mapped as \`string | null\` (via \`?? null\`), the \`!== null\` guard works. If mapped as \`string | undefined\` (via \`?? undefined\`), the guard silently passes \`undefined\` through. Fix: ensure \`parseDebugFile\` maps \`codeId\` to \`string | null\` (using \`obj.codeId ?? null\`) so \`check.ts\`'s existing \`!== null\` guard remains correct without touching \`check.ts\`.
 
@@ -365,9 +368,6 @@
 <!-- lore:019f0072-51a3-74b5-bb33-ccc06d2a01b1 -->
 * **skill-eval E2E tests fail on Anthropic API network errors — not a code regression**: Trap: \`test/e2e/skill-eval.test.ts\` failures (\`claude-sonnet-4-6 meets threshold\`, \`claude-opus-4-6 meets threshold\`) look like regressions introduced by the current PR. Root cause: these tests call \`api.anthropic.com\` directly — \`\[planner] API error: Invalid response body ... Premature close\` is an external Anthropic API outage, not a code bug. All 126 non-LLM E2E tests pass. Fix: confirm by checking logs for \`Premature close\` pattern; if present, stop re-running (wastes CI resources) and post a PR comment documenting the outage. Do not merge while CI is red — wait for API recovery.
 
-<!-- lore:019e471c-df3f-7e78-9df7-6b36d2d258db -->
-* **SQLite transaction() ROLLBACK can throw, discarding original error**: (gotcha) SQLite transaction ROLLBACK error-swallowing trap: In \`src/lib/db/sqlite.ts\`, \`transaction()\` catches errors and runs \`this.db.exec('ROLLBACK')\`. If ROLLBACK itself throws, the original error is lost. Fix: \`const origErr = e; try { this.db.exec('ROLLBACK'); } catch (rbErr) { log.debug(...); } throw origErr;\`
-
 <!-- lore:019f6b92-6ed7-71a9-8526-23f2f4e79289 -->
 * **Stale-lock staleness heuristic rests on factually wrong premise**: Trap: The stale-lock cleanup heuristic in sqlite.ts (lines 218-273) assumes lock directories are only modified at creation time, making mtime a reliable staleness signal. But Node.js modifies lock directories during normal operation (seen in strace), invalidating the premise. Fix: Remove the mtime check entirely — rely solely on the 60s timeout. The heuristic provides no actual protection against deleting live locks.
 
@@ -530,6 +530,9 @@
 * **wizard-runner.ts: large shared context via initialState, not inputData — D1 row size limit**: In \`wizard-runner.ts\`, large shared context (\`dirListing\`, \`fileCache\`, \`existingSentry\`) travels via \`initialState\` (not \`inputData\`) to avoid D1 per-row size overflow (see getsentry/cli-init-api#98). \`MAX\_RESUME\_RETRIES = 3\`, \`RETRY\_BACKOFF\_MS = \[2000, 4000, 8000]\`. \`resumeWithRetry()\` handles stale-step recovery via \`tryRecoverCurrentRunState()\` when \`isStepAlreadyAdvancedError()\` detects 'was not suspended' 500.
 
 ### Preference
+
+<!-- lore:019fd674-68f7-7e92-bcef-b0b8382e83e2 -->
+* **Always monitor CI checks to full completion before proceeding**: The user consistently waits for all CI checks on a PR to fully resolve (pass, fail, or skip) before taking next steps such as merging, promoting from draft, or addressing review comments. This applies across PRs regardless of size. The user expects the assistant to poll CI status until no checks are pending, report the complete list of check results (name, status, duration), and only then surface any failures or next actions. Never treat CI as done while any check is still in a pending/in-progress state.
 
 <!-- lore:019fd64a-7817-7223-a0f3-3610cb893d3a -->
 * **Always request recaps in ≤40 words, plain sentences, no markdown, leading with goal then next action**: When returning after stepping away, the user consistently requests a recap with strict constraints: ≤40 words, 1-2 plain sentences, no markdown formatting, lead with the current goal/task, end with the single next action. Skip root-cause narratives, fix internals, secondary to-dos, and em-dash tangents. Follow this format automatically whenever the user asks for a recap or summary after being away.
